Move curve.h to public include
