+++ include/ccf/crypto/curve.h
@@ -2,11 +2,12 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
+#include "ccf/crypto/md_type.h"
+#include "ccf/ds/enum_formatter.h"
 #include "ccf/ds/json.h"
-#include "ccf/ds/logger.h"
-#include "crypto/hash.h"
 
-#include <openssl/evp.h>
+#define FMT_HEADER_ONLY
+#include <fmt/format.h>
 #include <stdexcept>
 #include <string>
