+++ src/common/configuration.h
@@ -3,10 +3,10 @@
 
 #pragma once
 
+#include "ccf/crypto/curve.h"
 #include "ccf/ds/logger.h"
 #include "common/enclave_interface_types.h"
 #include "consensus/consensus_types.h"
-#include "crypto/curve.h"
 #include "ds/oversized.h"
 #include "ds/unit_strings.h"
 #include "enclave/consensus_type.h"
