+++ src/node/identity.h
@@ -2,8 +2,8 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
+#include "ccf/crypto/curve.h"
 #include "crypto/certs.h"
-#include "crypto/curve.h"
 #include "crypto/openssl/key_pair.h"
 #include "crypto/verifier.h"
